Extract login page into a fragment
